requirements.txt
APScheduler==3.5.3
argcomplete==1.9.4
Click==7.0
commentjson==0.7.1
cryptography==1.7.1
enum34==1.1.6
Flask==1.0.2
funcsigs==1.0.2
futures==3.2.0
idna==2.2
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.10
keyring==10.1
keyrings.alt==1.3
luma.core==1.8.2
luma.emulator==1.1.0
Adafruit_DHT==1.4.0
luma.lcd==1.1.0
luma.led-matrix==1.1.1
luma.oled==2.5.1
MarkupSafe==1.0
monotonic==1.5
Pillow==5.3.0
psutil==5.4.8
pyasn1==0.1.9
pycrypto==2.6.1
pygame==1.9.4
pygobject==3.22.0
pytz==2018.7
pyxdg==0.25
RPi.GPIO==0.6.4
SecretStorage==2.3.1
six==1.10.0
smbus2==0.2.1
sn3218==1.2.7
spidev==3.2
tzlocal==2.0.0b1
Werkzeug==0.14.1
